push

+++ main.py
@@ -17,7 +17,7 @@ def parse_arguments():
 
 def loss_fn(pred, gt):
     pred = SE3(pred)
-    gt = SE3(mat_to_quat(gt).to(device))
+    gt = SE3(mat_to_quat(gt).type(torch.float32).to(device))
 
     error = (gt.inv() * pred).inv()
     tr = error[:,0:3].norm(dim=-1)
